Client version/5.0 client (2018 12 25) (#1210)
* Основные изменения на 19 октября 2024 года:

[Game] Особенность: подключил базу версии 5.0.7.0;
- в базе английская локализация, поэтому в файле Config.json должна быть строка "DefaultLanguage": "en_us";

[Packets] Исправил структуры пакетов для лута;

[Login] Установил 4 открытых слота для создания персонажей;

[Game] Исправил слоты действий:
- в версии 5+ MaxActionSlots=157, вместо 133 для 3+;

[Game] Исправил работу баффа;

[Game] Автоматическая юстировка констант шифрации.

[Game] Исправлен бафф Dash:
- теперь мана расходуется при беге;
- используем формулу трата маны=3.33*уровень+11.67

[SQL] Добавлен файл loots_5070_add_memory-ember.sql:
- Исправляем ошибку: В камине нельзя взять уголёк Id=36917 "Memory Ember"

[SQL] Изменил поле `slot_type` с enum на int:
- добавлен файл 2024-09-16_aaemu_game_item_containers_update.sql;
- добавлен файл 2024-09-16_aaemu_game_items_update.sql;
- обновлен файл aaemu_game.sql.

[Game] Внесены изменения в код:
- Теперь работает Divine Clock;
- Получаем подарки за онлайн в игре.
- Теперь подарки, которые будут вручены выбираются из расписания;

[Game] Исправлен домашний портал:
- Можно переименовать/удалить домашний портал;

[Data] Исправил точки крепления для домов (сгенерировано плагином HousingBindings);

[SQL] Добавлен файл 2024-06-09_aaemu_game_divine_clock.sql:
- Таблица 'divine_clock' осуществляет поддержку выдачи подарков за нахождение в онлайн в игре;

[Game] Исправил возможность покинуть транспорт;

[Game] Исправил возможность переименовать питомцев, транспорт и дома;

[Game] WIP: начато исправление аукциона;
- Сохранение в базу предметов аукциона - не реализовано;

Исправление поиска предметов:
- постраничное листание списка лотов;
- Добавлен определитель языка для keyword.
- Для правильного поиска по аукциону потребуется добавить переводы в серверную базу, иначе не будет поиска на родном языке.
- Пока поиск будет только на английском языке;

Продажа:
- показывает только список лотов, которые игрок выставил на продажу;

Ставки:
- Можно делать ставки;
- Отображается в каких лотах мы сделали ставку;
- Добавлен вывод только наших ставок.

- Для целей проверки ставка вместо 6 часов будет продолжаться 6 минут;
- Добавлен частичный выкуп лота;

[SQL] обновлены aaemu_game.sql и добавлен 2024-10-04_aaemu_game_auction_house_update.sql

[Game] WIP: начато исправление группы;

[Game] Исправление инициализации инвентаря;
[Game] Исправлено назначение владельца предмета при разделении стопки на две части;

[Game] Исправление почты:
- исправил принятие почты;
- теперь почта не пропадает при закрытии письма;
- при получении прикрепленных предметов к письму, пустое письмо можно сразу удалить (активируется кнопка удаления);

[Game] Исправление покупки/продажи у Npc;

[Game] Исправил возможность ставить/убирать оборудование на корабли;

* Основные изменения на 29 октября 2024 года:

[Game] WIP: исправление аукциона:
- Добавлено сохранение в базу данных;

[Game] Добавлены перечисления MailBoxListKind, SystemMailSenderKind;
- перемещены все `enum` файлы в Mail\Static

[Game] WIP: исправление аукциона:
- Исправил покупку части аукционного лота;

[SQL] updated aaemu_game.sql and added 2024-10-25_aaemu_game_auction_house_update.sql

[Game] WIP: исправление аукциона:
- добавил расчёт и сохранение динамики продаж;

[SQL] updated aaemu_game.sql and added 2024-10-28_aaemu_game_auction_solds_data_update.sql

+++ AAEmu.Game/Core/Packets/G2C/SCScheduleItemUpdatePacket.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+
 using AAEmu.Commons.Network;
 using AAEmu.Game.Core.Network.Game;
 using AAEmu.Game.Models.Game;
@@ -16,7 +17,7 @@ public class SCScheduleItemUpdatePacket : GamePacket
 
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(_scheduleItems.Count);
+        stream.Write((byte)_scheduleItems.Count);
         foreach (var item in _scheduleItems)
         {
             stream.Write(item);
